requirements.txt
fastapi
uvicorn
motor
pydantic-settings
python-dotenv
discord.py
google-generativeai
aiohttp
